Simplify docker image handling
The base image makes the build process more efficient, as the expensive
initial setup can be reused. We do, however, have to run this manually,
which may be forgotten or trigger ill-timed problems.
The caching built into the docker build and push action solves this
problem in a transparent number, yielding similar or even faster build
times to the current approach.