Add workflows for cluster deployment and testing
The API test and bare metal deploy workflows are reusable, i.e. they
can easily be reused, e.g. to apply the tests to different deployment
methods, or to have different triggers for the deployment. For now, the
deployment and test is triggered manually, but this can be extended to
deploy against a dedicated cluster on pull requests.

Note that this requires an additional workflow, as well as a decision on
labels for self-hosted runners, as the e2e workflows also use this
mechanism.