Fix errors on stopping containers
Removing containers by name is a repeating pattern that fails when the
container in question is not running. This sometimes happens during
execution. Since we want to stop the container anyways, we can ignore
this. The pattern is replaced by a helper that consistently handles
removing a set of containers by name.