Pci driver handling (#613)
* Remove unused utils

* Unify handling of pci devices

* Fix SFAM-2235 (#542)

* Introduce helper to check PCI driver presence

* Attempt all supported PCI drivers for SPDK devices

Previously only `uio_pci_generic` was attempted. This resulted in
failure if the driver was not loaded. This guards against the failure,
attempts `vfio-pci`, and only fails with a descriptive error if neither
is available.

* Fix SPDK PCI driver handling (#543)

According to the DPDK documentation, vfio-pci is the preferred driver.
The vfio-module needs to be parametrized to allow binding devices without
an iommu group though. As a fallback uio_pci_generic is used instead.

* added second check for enable_unsafe_noiommu_mode (#567)

* added second check for enable_unsafe_noiommu_mode

* use vfio-pci if device has an iommu group

* removed file @

* fixed Undefined name snode_ops

* to revert: update env_var

* Fix failing device list on unloaded driver

* to revert: update version

* reverted env_var changes

---------

Co-authored-by: Max Schettler <max@simplyblock.io>
Co-authored-by: Manohar Reddy <manohar@simplyblock.io>